refactor(domain) : match 패키지 수정 및 전 entity 정리(@setter, @builder, @AllArgsConstructor 제거)
