+++ src/main/java/com/likelion/zooting/domain/match/entity/Match.java
@@ -6,9 +6,7 @@ import lombok.*;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)  // JPA만 접근 가능
-@AllArgsConstructor
 @Table(name = "MATCH")
 public class Match {
     @Id
