+++ src/main/java/com/likelion/zooting/domain/match/policy/MatchScoreType.java
@@ -1,4 +1,4 @@
-package com.likelion.zooting.domain.match.service;
+package com.likelion.zooting.domain.match.policy;
 
 import com.likelion.zooting.domain.match.exception.MatchInnerErrorCode;
