+++ src/main/java/com/likelion/zooting/domain/user/entity/User.java
@@ -8,10 +8,7 @@ import java.time.LocalDateTime;
 
 @Entity
 @Getter
-@Setter
-@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)  // JPA만 접근 가능
-@AllArgsConstructor
 @Table(name = "USER")
 public class User {
     @Id
