+++ src/main/java/com/likelion/zooting/domain/userinterest/entity/UserInterest.java
@@ -7,9 +7,7 @@ import lombok.*;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)  // JPA만 접근 가능
-@AllArgsConstructor
 @Table(name = "USER_INTEREST")
 public class UserInterest {
     @Id
