+++ src/main/java/com/likelion/zooting/domain/userpreferredanimaltype/entity/UserPreferredAnimalType.java
@@ -7,9 +7,7 @@ import lombok.*;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)  // JPA만 접근 가능
-@AllArgsConstructor
 @Table(name = "USER_ANIMAL_TYPE")
 public class UserPreferredAnimalType {
     @Id
